Resolves #12 - Now handles capture, passive and once or event handlers. Prevents passive and auto-calling preventDefault() Resolves #28 - New logo included. One for on dark background and one for on light backgrounds Resolves #48 - Added a way to define additional element tags that should not be minified. Resolves #61 - Added a filter to the router.onUpdate method Resolves #62 - We now pass in relevant information about the current route. Resolves #65 - EvoRouter now properly handles trailing / Resolves #66 - All the source files are not ESM based. Resolves $67 - Comments for a looped variable name are now accurate Resolves #68 - Now looped Event Handlers use the correct variable names Resolves #71- We now pass in relevant information about the current route into the public method navParams if it exists.
Added watch to the demo/test app
Updates to README.md
Changed from tsconfig.json to jsconfig.json since we use JSDOC types instead of TypeScript.
Cleaned up the "selected" feature of the <wc-a> element
Updated the RouteNotFoundElement.html file to include an SVG image.
Added lost of comment from conversations with chatGPT on various improvements to the routing and dialogs.
Added ability to remove/delete elements that are router targets when the developer adds the attribute `auto-delete` or `auto-delete="<minutes>"` to `<wc-route>` to help save memory
Added parseQueryParams to EvoRouter.js file to convert query params into an object.
Improved `Camel to Kebab` and `Kebab to Camel` routines to better handle i18n words.
Lots of spelling corrections
Improved support for for loops in preparation to allow sub loops.
Added router support for hashchange events